Added exit to deleteTask in UI

+++ UI.cs
@@ -95,13 +95,13 @@ while (true)
     void DeleteTaskUI()
     {
         bool correctDeletion = false;
-        int deleteSelection = -1;
+        int deleteSelection = -2;
 
         while (!correctDeletion)
         {
 
             Console.Clear();
-            Console.WriteLine("Please enter the number of the task you wish to delete.");
+            Console.WriteLine("Please enter the number of the task you wish to delete. To cancel, enter [-1]");
             Console.WriteLine();
 
             int x = 0;
@@ -111,26 +111,38 @@ while (true)
                 x++;
             }
 
-            while (deleteSelection == -1)
+            while (deleteSelection == -2)
             {
                 try
                 {
-                    deleteSelection = int.Parse(Console.ReadKey(true).KeyChar.ToString());
+                    deleteSelection = int.Parse(Console.ReadLine());
                 }
                 catch { }
 
+                if (deleteSelection == -1)
+                {
+                    break;
+                }
                 if (deleteSelection < 0 || deleteSelection > currentUser.Tasks.Count)
-                    deleteSelection = -1;
+                    deleteSelection = -2;
             }
 
-            Console.WriteLine($"Deleting task {deleteSelection}: {currentUser.Tasks[deleteSelection].Title}, is this correct? (Y/N)");
-
-            string deletionConfirm = Console.ReadKey(true).KeyChar.ToString();
-
-            if (deletionConfirm.ToLower() == "y")
+            if (deleteSelection == -1)
                 correctDeletion = true;
-        }
 
-        currentUser.DeleteTaskByIndex(deleteSelection);
+            if (deleteSelection != -1)
+            {
+
+                Console.WriteLine($"Deleting task {deleteSelection}: {currentUser.Tasks[deleteSelection].Title}, is this correct? (Y/N)");
+
+                string deletionConfirm = Console.ReadKey(true).KeyChar.ToString();
+
+                if (deletionConfirm.ToLower() == "y")
+                    correctDeletion = true;
+
+                currentUser.DeleteTaskByIndex(deleteSelection);
+            }
+
+        }
     }
 }
